Move out score checking function
